+++ pages/search/index.tsx
@@ -15,9 +15,9 @@ import RecipeList from "@/components/searchPage/RecipeList";
 import ChangeSortedBtn from "@/components/searchPage/ChangeSortedBtn";
 import SideFoodCate from "@/components/searchPage/SideFoodCate";
 import SideCookingTime from "@/components/searchPage/SideCookingTime";
-import { FieldErrors, useForm } from "react-hook-form";
 import TopButton from "@/components/button/TopButton";
 import { GrRotateLeft } from "react-icons/gr";
+import SearchTextBar from "@/components/searchPage/SearchTextBar";
 
 const SearchData: NextPage = () => {
     const router = useRouter();
@@ -30,15 +30,6 @@ const SearchData: NextPage = () => {
     const [totalItems, setTotalItems] = useState<TypeRecipe[]>([]);
     const [lastDoc, setLastdoc] = useState(0);
 
-    const { register, handleSubmit, getValues } = useForm();
-    const onValid = () => {
-        sessionStorage.setItem("searchData", getValues("searchText"));
-        setText(getValues("searchText"));
-    };
-    const onInValid = (errors: FieldErrors) => {
-        console.log(errors);
-    };
-
     // 인기순
     const activeBestBtn = () => {
         sessionStorage.setItem("userWatching", "viewCount");
@@ -50,7 +41,7 @@ const SearchData: NextPage = () => {
         sessionStorage.setItem("userWatching", "createdAt");
         setIsBest("createdAt");
     };
-    // 전체목록(6개씩)
+    // 전체목록(8개씩)
     const first = async () => {
         const querySnapshot = await getDocs(
             query(
@@ -59,7 +50,7 @@ const SearchData: NextPage = () => {
                     isBest === "viewCount" ? "viewCount" : "createdAt",
                     "desc"
                 ),
-                limit(6)
+                limit(8)
             )
         );
         const newData = querySnapshot.docs.map((doc: any) => ({
@@ -89,7 +80,7 @@ const SearchData: NextPage = () => {
                     "desc"
                 ),
                 startAfter(lastDoc),
-                limit(6)
+                limit(8)
             )
         );
         updateState(querySnapshot);
@@ -109,7 +100,7 @@ const SearchData: NextPage = () => {
     };
     // 검색
     const fuse = new Fuse(currentItems, {
-        keys: ["animationTitle", "foodTitle"],
+        keys: ["animationTitle", "foodTitle", "cookingTime"],
         includeScore: true,
         threshold: 0.5, //일치정도(0~1.0)
         minMatchCharLength: text.length,
@@ -176,8 +167,8 @@ const SearchData: NextPage = () => {
         const storeFilteredTime = JSON.parse(
             sessionStorage.getItem("filteredTimeData")!
         );
-        result ? setIsBest(result) : setIsBest("createdAt");
 
+        result && setIsBest(result);
         storeSearchText && setText(storeSearchText);
         storeFilteredFood && setFilteredFood(storeFilteredFood);
         storeFilteredTime && setFilteredTime(storeFilteredTime);
@@ -189,38 +180,10 @@ const SearchData: NextPage = () => {
         <>
             <div className="w-full mt-20 flex flex-col justify-center items-center">
                 <TopButton />
-                <form
-                    onSubmit={handleSubmit(onValid, onInValid)}
-                    className="relative mt-4 mb-16 flex"
-                >
-                    <input
-                        {...register("searchText")}
-                        type="text"
-                        className="w-[300px] h-[50px] text-sm font-medium pl-7 focus:outline-none rounded-sm rounded-r-none border border-slate-300"
-                        placeholder="하울의 움직이는 성 베이컨계란요리"
-                    ></input>
-                    <button
-                        type="submit"
-                        className="bg-brand100 rounded-sm rounded-l-none w-[50px] h-[50px] text-center"
-                    >
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            fill="none"
-                            viewBox="0 0 24 24"
-                            strokeWidth="1.5"
-                            stroke="currentColor"
-                            className="w-6 h-6 text-white absolute top-3 ml-3 pointer-events-none"
-                        >
-                            <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607z"
-                            />
-                        </svg>
-                    </button>
-                </form>
+                <SearchTextBar setText={setText} />
                 <ChangeSortedBtn
                     text={text}
+                    setText={setText}
                     currentItems={currentItems}
                     dataResults={dataResults}
                     isBest={isBest}
@@ -249,7 +212,7 @@ const SearchData: NextPage = () => {
                             선택초기화<GrRotateLeft></GrRotateLeft>
                         </button>
                     </div>
-                    <div className="grid mx-auto sm:mx-0 sm:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-3 gap-x-7 gap-y-9 relative pb-24">
+                    <div className="grid mx-auto sm:mx-0 sm:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-4 gap-x-7 gap-y-9 relative pb-24">
                         <RecipeList
                             text={text}
                             next={next}
